requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.14
django-bootstrap-v5==1.0.11
paho-mqtt==2.1.0
soupsieve==2.5
sqlparse==0.5.1
tzdata==2024.1